+++ bin/Cargo.toml
@@ -97,6 +97,7 @@ once_cell = "1.10"
 serial_test = "0.8"
 
 [target.'cfg(target_os="linux")'.dev-dependencies]
+pnet = "0.28"
 systemd = { package = "systemd", version = "0.7" }
 
 [build-dependencies]
