+++ bin/tests/it/main.rs
@@ -2,6 +2,8 @@ mod cli;
 mod cmd_line_env;
 mod common;
 mod http;
-mod k8s;
 mod metrics;
 mod retries;
+
+#[cfg(target_os="linux")]
+mod k8s;
